(BSR)[API] chore: Align PostVenueBodyModel and PatchOfferBodyModel
serializers in the way they handle address data. The frontend will
use the same component for autocompletion/manual edition so it make
sense that both of thoses routes expect same data structure
